refactor: UtilFunctions and remove NotificationPermissionHelper
- Updated UtilFunctions.kt to improve code readability and added isNetworkAvailable function to check network connectivity.
- Removed NotificationPermissionHelper.kt as it was no longer needed.
- Updated build.gradle to use newer versions of Gradle and dependencies.
- Changed android.enableJetifier to false in gradle.properties for better compatibility.
- Updated gradle-wrapper.properties to use Gradle version 8.13.
- Added ConnectScreenComponents.kt to implement UI components for the connect screen, including tabs, search functionality, and friend/circle cards.
