+++ app/build.gradle
@@ -5,8 +5,8 @@ plugins {
     id 'androidx.navigation.safeargs.kotlin'
     id 'com.google.gms.google-services'
     id 'com.google.firebase.crashlytics'
-    id 'org.jlleitschuh.gradle.ktlint' version '12.1.0'
-    id 'org.jetbrains.kotlin.plugin.compose' version '2.1.21'
+    id 'org.jlleitschuh.gradle.ktlint' version '13.0.0'
+    id 'org.jetbrains.kotlin.plugin.compose' version '2.2.0'
 }
 
 def keystorePropertiesFile = rootProject.file("keystore.properties")
@@ -19,7 +19,7 @@ if (keystorePropertiesFile.exists()) {
 }
 
 android {
-    compileSdk 35
+    compileSdk 36
 
     signingConfigs {
         release {
@@ -34,7 +34,7 @@ android {
     defaultConfig {
         namespace "com.dscvit.vitty"
         minSdkVersion 26
-        targetSdkVersion 35
+        targetSdkVersion 36
         versionCode 40
         versionName "2.0.3"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -61,7 +61,7 @@ android {
 
 dependencies {
     // Compose BOM (Bill of Materials)
-    implementation platform("androidx.compose:compose-bom:2025.06.01")
+    implementation platform("androidx.compose:compose-bom:2025.07.00")
 
     // Core Compose dependencies
     implementation "androidx.compose.ui:ui"
@@ -78,12 +78,12 @@ dependencies {
     implementation 'com.google.android.material:material:1.12.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.2.1'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
-    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.9.1'
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.9.1'
+    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.9.2'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.9.2'
     implementation 'androidx.preference:preference-ktx:1.2.1'
-    implementation 'androidx.navigation:navigation-fragment-ktx:2.9.1'
-    implementation 'androidx.navigation:navigation-ui-ktx:2.9.1'
-    implementation 'androidx.navigation:navigation-compose:2.9.1'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.9.2'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.9.2'
+    implementation 'androidx.navigation:navigation-compose:2.9.2'
     implementation 'androidx.compose.material3:material3-android:1.3.2'
     implementation 'androidx.compose.runtime:runtime-livedata:1.8.3'
     testImplementation 'junit:junit:4.13.2'
@@ -112,18 +112,18 @@ dependencies {
     releaseImplementation 'com.google.firebase:firebase-crashlytics-ktx'
 
     // Retrofit
-    implementation 'com.squareup.retrofit2:retrofit:2.9.0'
-    implementation 'com.squareup.okhttp3:logging-interceptor:4.7.2'
-    implementation 'com.squareup.retrofit2:converter-gson:2.9.0'
+    implementation 'com.squareup.retrofit2:retrofit:3.0.0'
+    implementation 'com.squareup.okhttp3:logging-interceptor:5.1.0'
+    implementation 'com.squareup.retrofit2:converter-gson:3.0.0'
 
     // Loading Animations
     implementation 'com.github.ybq:Android-SpinKit:1.4.0'
 
     // Coil
-    implementation "io.coil-kt:coil:2.6.0"
-    implementation("io.coil-kt:coil-compose:2.6.0")
+    implementation "io.coil-kt:coil:2.7.0"
+    implementation("io.coil-kt:coil-compose:2.7.0")
     // Timber for logs
-    implementation 'com.jakewharton.timber:timber:4.7.1'
+    implementation 'com.jakewharton.timber:timber:5.0.1'
 
     // Room Database
     implementation "androidx.room:room-runtime:2.7.2"
@@ -131,8 +131,8 @@ dependencies {
     kapt "androidx.room:room-compiler:2.7.2"
 
     // ViewModel and LiveData
-    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.9.1"
-    implementation "androidx.lifecycle:lifecycle-runtime-compose:2.9.1"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-compose:2.9.2"
+    implementation "androidx.lifecycle:lifecycle-runtime-compose:2.9.2"
 
     implementation 'androidx.dynamicanimation:dynamicanimation-ktx:1.1.0'
     implementation 'com.github.jeziellago:compose-markdown:0.5.7'
