+++ build.gradle
@@ -7,11 +7,11 @@ buildscript {
         maven { url "https://jitpack.io" }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:8.10.1'
+        classpath 'com.android.tools.build:gradle:8.11.1'
         classpath 'org.jetbrains.kotlin:kotlin-gradle-plugin:2.1.21'
-        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.8.4'
-        classpath 'com.google.gms:google-services:4.4.2'
-        classpath 'com.google.firebase:firebase-crashlytics-gradle:3.0.2'
+        classpath 'androidx.navigation:navigation-safe-args-gradle-plugin:2.9.2'
+        classpath 'com.google.gms:google-services:4.4.3'
+        classpath 'com.google.firebase:firebase-crashlytics-gradle:3.0.4'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
